+++ router/admin/contact.js
@@ -3,7 +3,7 @@ const router = express.Router();
 const contact = require('../../controller/admin/contact.js')
 
 router.get('/', contact.getContact)
-router.post('/', contact.postContact)
+router.get('/view/:ID', contact.viewContact)
 router.get('/delete/:ID', contact.deletContact)
 
 module.exports = router;
