+++ views/dashboard/contact.ejs
@@ -38,7 +38,7 @@
                 <div class="col-xl-8 col-lg-6">
                     <div class="contact-form">
                         <h2 class="h2-title">it would be great to hear from you</h2>
-                        <form action="/admin/contact" method="post" class="form">
+                        <form action="/contact" method="post" class="form">
                             <div class="row">
                                 <div class="col-12">
                                     <input type="text" class="form-input" placeholder="Full Name" name="name">
